requirements.txt
# requirements.txt
# Unk Agent - Production Dependencies
# Who Visions LLC

# ═══════════════════════════════════════════════════════════════════════════
# CORE FRAMEWORK
# ═══════════════════════════════════════════════════════════════════════════
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-multipart>=0.0.12

# ═══════════════════════════════════════════════════════════════════════════
# GOOGLE CLOUD / GEMINI
# ═══════════════════════════════════════════════════════════════════════════
google-genai>=0.6.0
google-cloud-firestore>=2.19.0
google-cloud-logging>=3.11.0
google-cloud-secret-manager>=2.21.0

# ═══════════════════════════════════════════════════════════════════════════
# AUTHENTICATION
# ═══════════════════════════════════════════════════════════════════════════
firebase-admin>=6.5.0
python-jose[cryptography]>=3.3.0

# ═══════════════════════════════════════════════════════════════════════════
# ASYNC / CONCURRENCY
# ═══════════════════════════════════════════════════════════════════════════
httpx>=0.27.0
aiohttp>=3.10.0

# ═══════════════════════════════════════════════════════════════
# UTILITIES
# ═══════════════════════════════════════════════════════════════
python-dotenv>=1.0.0
tenacity>=9.0.0

# ═══════════════════════════════════════════════════════════════
# SCRAPING
# ═══════════════════════════════════════════════════════════════
beautifulsoup4>=4.12.0
requests>=2.31.0
